Private API testing done. hacked market order
The test framework for the private API is complete. It is not possible
to get fills in the testing environment because the orders do not
execute. So, we cannot test that API there. I am receiving Orders with
no price because of the new market orders, so had to hack FromJSON
Order instance to allow that for now. Needs a fix.