Enable BTC transfers to an arbitrary wallet
FIX ME! This fails if we try to transfer amounts smaller than 0.1 BTC
(e.g. 0.01 as that is `shown` as "1.0e-2" and does not pass
validation). The proper way to fix it is to change the ToJSON instance
of CoinScientific, but I'm not messing with that at this point as I
believe volumes should not be expressed in terms of CoinScientific, but
in `Volume`. This problem will also be fixed by that change.