FIX gdax `done` msg parsing & Add heartbeat
This is the final for the previous “work around”.

GDAX no longer provides an “order_type” field on their `done` websocket
messages. This was also failing the parser. This is now fixed. I also
added the Heartbeat messages to the message type. But there are no
tests for it yet!