Revert & Fix typeclass Instances
Fixing comparisons to 2 decimal places for prices and 8 d.p. for sizes
(is very convenient but) works only for the BTC-USD market. In other
markets, such as the ETH-BTC market, we need more precision on prices
and sizes. So, I have to rollback these changes. I am also changing the
ToJSON and Show instances to show whatever precision we have. If we ask
for too much precision, order placements will fail. We will need to
take care of that before we make any such request.